Refactor code into profile data context and pass into index.js
